Fix AppListViewScalableLayoutTest for productivity launcher
The tests were testing apps container layout for legacy launcher - have
the existing tests explicitly disable productivity launcher feature.

Adds new set of tests that perform similar layout verification for
productivity launcher apps container (used in tablet mode), both for
cases recent apps are shown, and case recent apps are not shown.

Fixes few issues that surfaced:
*   new spec does not reduce search box height to 40 dip on screen size
    < 600.
*   change how min horizontal margin is calculated for productivity
    launcher - it's a constant set to 24
*   fix logic for calculating number of rows when the row count in the
    grid needs to increased - the logic required std::ceil of a result
    of division, but it was performing an integer division, which
    yielded incorrect result - add a static cast to float to address
    this.

BUG=1280300

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3423448
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#965723}

+++ ash/app_list/views/paged_apps_grid_view.cc
@@ -1382,9 +1382,9 @@ int PagedAppsGridView::CalculateMaxRows(int available_height,
     // I.e. min n, with padding as close to max as possible where:
     // n* tile_height + (n - 1) * padding <= available_height
     // padding <= max_padding
-    final_row_count =
-        std::ceil((available_height + kMaxVerticalPaddingBetweenTiles) /
-                  (tile_height + kMaxVerticalPaddingBetweenTiles));
+    final_row_count = std::ceil(
+        static_cast<float>(available_height + kMaxVerticalPaddingBetweenTiles) /
+        (tile_height + kMaxVerticalPaddingBetweenTiles));
   }
   // Unit tests may create artificially small screens resulting in
   // `final_row_count` of 0. Return 1 row to avoid divide-by-zero in layout.
